requirements.txt
bencode.py>=4.0.0
isbnlib>=3.10.14
tqdm>=4.67.1
zstandard>=0.23.0
pillow>=11.1.0
docopt-ng>=0.9.0
pyvips>=2.2
microjson>=0.4.1